requirements.txt
openai>=1.0.0
newsapi-python
fredapi
pandas
numpy
tqdm
ollama 
